pontevpn/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
stripe==9.9.0
resend==1.0.1
redis==5.0.4
slowapi==0.1.9
paramiko==3.4.0
python-dotenv==1.0.1
httpx==0.27.0
